local commit
apod generator works, users can enter a value between 6 and 100, which corresponds to the number of images randomly generated and fed back to the frontend. Attempting a search bar to allow users to search for apods, however I'm unsure on how to do it; need to speak to tom to see if he have any ideas.