local commit
added a historical section on the apod page, as well as splitting all information up into neat tables. Sections now consist of: daily apod which cant be changed by user, historic apod which uses the ${choice} to let the user choose an apod, formerly just the apod section, and the apod generator.